% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KTU2.R
\name{kaxaDB}
\alias{kaxaDB}
\title{Pre-trained 16S Databases for Kaxonomy annotation}
\format{
Pre-trained tetranucleotide database and taxonomy information (path to RDS files) of 16S rRNA genes.
}
\description{
The pre-trained databases contain a species+strain levels 16S rRNA gene database retrieved from NCBI (2022/02/10) and SILVA 7-level 16S rRNA gene database (ver. 138).
 Two hypervariable-region trimmed formats, V3V4 (341F-805R, primer-trimmed) and V4 (515F-806R, primer-trimmed) regions, are available.
 A pair of files (DB and TX, RDS format) are required for annotation by using 'kaxonomy' function.
 See how to use a pre-trained database for taxonomy annotation in Examples.
Available built-in databases: NCBI_V3V4, NCBI_V4, SILVA138_V3V4, and SILVA138_V4
}
\examples{

kluster <- readRDS("kluster.RDS") # read klustering/ktusp result object from RDS file
kaxa <- kaxonomy(dbRDS = "NCBI_V3V4",
                 taxaRDS = NULL,
                 kmer.table = kluster$kmer.table,
                 cores = 4)

write.table(kaxa,file = "kaxonomy.tsv",sep="\t") # write out with tab-delimited text file
saveRDS(kaxa,"kaxonomy.RDS") # write out with RDS format
}
