% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KTU2.R
\docType{data}
\name{phyloseq_marsh}
\alias{phyloseq_marsh}
\title{Subsampled phyloseq object from metabarcoding of 16S from marsh soils}
\format{
A phyloseq object containing 2417 ASVs and 10 samples.
}
\source{
\url{https://github.com/csmiguel/marsh_metabarcoding}
}
\usage{
data(phyloseq_marsh)
}
\description{
This dataset contains subsampled most frequent ASVs and their counts as a matrix across 41 samples
 from marsh soils in southern Spain.
 They come from: Camacho-Sanchez et al. Bacterial assemblage in Mediterranean salt marshes: disentangling
 the relative importance of seasonality, zonation and halophytes. In review.
}
\keyword{datasets}
