% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KTU2.R
\name{ps2KTU}
\alias{ps2KTU}
\title{Klustering using as input a phyloseq object}
\usage{
ps2KTU(
  ps = NULL,
  subset = NULL,
  path2fasta = NULL,
  dnaseqs = NULL,
  method = "klustering",
  ...
)
}
\arguments{
\item{ps}{phyloseq object.}

\item{subset}{(optional) 'numeric' for debugging. Limit klustering to a randomly sampled subset of ASVs/OTUs.}

\item{dnaseqs}{otu_table/refseq: slot to search for DNA sequences in phyloseq; either as row/col names in otu_table or as a DNAStringSet in the refseq.}

\item{method}{either 'klustering' or 'ktusp'.}
}
\value{
KTU.table Aggregated KTU table.

ReqSeq Representative KTU sequences

kmer.table Tetranucleotide present score table or tetranucleotide frequency table.

clusters K-clusters of input features.
}
\description{
It reads a phyloseq object with the refseq slot.
}
\examples{
ps <- readRDS(
url("https://zenodo.org/record/4155877/files/Schreuder2020_TemporalDynamicsCloacalMicrobiota_phyloseq.rds"))
# run klustering on a subset of the ASVs with DNA sequences as ASV name in otu_table:
ps2KTU(ps = ps, subset = 200, dnaseqs = "otu_table")
# run klustering on a subset of the ASVs with DNA sequences in refseq slot:
data(phyloseq_marsh)
ps2KTU(ps = phyloseq_marsh, subset = 200, dnaseqs = "refseq")
# run ktusp:
ps2KTU(ps = ps, method = "ktusp", subset = 800, cores = 2, dnaseqs = "otu_table",
         split_tree_init = 5, split_lwrlim = 10000, split_reassemble = 200)
}
