% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KTU2.R
\docType{data}
\name{seqtab}
\alias{seqtab}
\title{Subsampled ASVs abundance table from metabarcoding of 16S from marsh soils}
\format{
A matrix containing 1000 columns, ASVS, and 41 rows, samples.
}
\source{
\url{https://github.com/csmiguel/marsh_metabarcoding}
}
\usage{
data(seqtab)
}
\description{
This dataset contains subsampled most frequent ASVs and their counts as a matrix across 41 samples
 from marsh soils in southern Spain.
 They come from: Camacho-Sanchez et al. Bacterial assemblage in Mediterranean salt marshes: disentangling
 the relative importance of seasonality, zonation and halophytes. In review.
}
\keyword{datasets}
