% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KTU2.R
\name{split_tetra}
\alias{split_tetra}
\title{Subset large tetranucleotide frequency table}
\usage{
split_tetra(
  repseq,
  split_tree_init = 5,
  split_lwrlim = 10000,
  split_reassemble = 1000,
  cores = 1
)
}
\arguments{
\item{repseq}{The fasta file path}

\item{split_tree_init}{Initial numbers for cutting kmer tree, default is 5}

\item{split_lwrlim}{Split kmer subtrees if tree tips > n, n default is 10000}

\item{split_reassemble}{Reassemble tree tips when subtrees are too small, default is 1000}

\item{cores}{Numbers of CPUs}
}
\value{
Splitted tetranucleotide frequency tables
}
\description{
Subset large tetranucleotide frequency table by splitting kmer tree
}
