% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KTU2.R
\name{tetra.freq}
\alias{tetra.freq}
\title{Generate a kmer table from representative sequences}
\usage{
tetra.freq(repseq, pscore = FALSE, file = TRUE, output.seq = FALSE, cores = 1)
}
\arguments{
\item{repseq}{The fasta file path}

\item{file}{logical, default is TRUE, loading DNA sequences from file. if 'file=FALSE', the 'repseq' is loaded from object.}

\item{output.seq}{logical, default if FALSE, outputting data of tetranucleotide frequency table and sequences.}

\item{cores}{Numbers of CPUs}
}
\value{
A 256 tetranucleotide combinations frequency table
}
\description{
Convert DNA sequences to tetranucleotide frequency table
}
