% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KTU2.R
\name{trim.primer}
\alias{trim.primer}
\title{Prepare KTU taxonomy assignment database by trimming primers}
\usage{
trim.primer(
  fasta,
  forseq,
  revseq,
  truncFL = NULL,
  output.name = NULL,
  progress = TRUE
)
}
\arguments{
\item{fasta}{input database fasta file}

\item{forseq}{forward primer sequence (5' -> 3')}

\item{revseq}{reverse primer sequence (5' -> 3')}

\item{truncFL}{trim n-bp length of the sequences from forward primer (for single-end analysis studies)}

\item{output.name}{(optional) the name for output file}

\item{progress}{show the processing progress}
}
\description{
Trim primer sequences of full length 16S/18S/target gene database by PCR primers
}
